tests/torch_datasets/sample/ refactor
